removed a warning

+++ src/api/fileimpl.cpp
@@ -108,8 +108,8 @@ FileImpl::getFileUri() const
     std::stringstream lEncodedResult;
     lEncodedResult << "file:///";
 
-    int lCurrentPos = 0;
-    int lNextSlashPos = lPath.find_first_of("/");
+    size_t lCurrentPos = 0;
+    size_t lNextSlashPos = lPath.find_first_of("/");
     while (lNextSlashPos != std::string::npos) {
       String lEncodedSegment(lPath.substr(lCurrentPos, lNextSlashPos - lCurrentPos));
       lEncodedResult << lEncodedSegment.encodeForUri() << "/";
